Changed behavior of retry_count option for mirrors
Now if query attempt to remote fails, it will reselect new mirror
according to current ha-strategy, and retry with this new selected
host. Earlier selecting of a mirror was only before the query, and
in case of fail sphinx could just try to connect again to the same
host several times.

Test added, documented.