Type marks in StringPtr for listing PQ queries
Percolate indexes may be either ql, either json queries. This is managed
by boolean flag stored with query. However when sending over network
we have strictly typed rows with only one type per resultset. So, having
differently typed queries in one index leads to non-uniform resultset we
need to deal with.

While it is not the problem for sql output (excepting the fact that two
different by type queries might be displayed as exactly equal), this is
more critical when returning listing in json (they have different
schema there depending of type).

For remote agents we use same code as for usual `select` statement, it
allows to all the things like filters/expressions work as for any other
resultsets, and since this we need to keep the proto untouched to avoid
huge code duplication and messing.

So, collecting listings from remote agents we need to keep someway type
info for every row in resultset keeping protocol unchanged and compatible
with usual usecases where no type required.

By discussion it is decided to solve the task by adding special mark at
the end of the string attr. Prerequisites are:
 1. Strings mainly used az z-terminated
 2. They send as arrays of bytes (with explicitly defined length)

So, we can use tail between zero terminator and actual capacity of string
to store additional non-displayed info (exactly one byte: 1 means 'ql',
0 means 'json'). This mark is kept in array, it is ignored when output
text via sphinxql, but used for json rendering.

FIXME! That is legacy, m.b. we need to consider pq indexes to be typed
as a whole; i.e. either index of ql queries, either index of jsons,
without keeping that type on per-query basis. It also will solve problem
with messing of displaying queries in ql, and will allow even to store
such queries in json form via sphinxql. Keeping mixed queries on one
index looks not a good and reasonable solution.