Downgrade and rename VER_MASTER
This constant was historically introduced only to manage master-agent
communitations especially in SEARCH command. But because of too generic
name it became confused with some kind of 'general' master version,
which leads sometime to mistakenly raising this num when introducing
new API commands, and breaking searches because of this.

It fixes #726 where such confusion caused incompatibility with
another branch of code, being run current 'master' branch as distr
index, and another as agents.