Revert wsrep sources to original state
Let's use unchanged wsrep sources instead of patched one.
That is commit 2c211e1 from github.com/percona/wsrep-API.git
This very source will be used both for building manticore and also for
building galera, so it is better to have them unchanged. The project
provides also CMakeLists.txt, and we use it to build wsrep in
subfolder and link with it.

Strictly speaking it would be better to just skip everything about
replication on platforms which not have galera, but currently it is
included in many places, so instead we just remove wsrep via a stub
header.