Refactor build with Galera
Previous variant uses in-source build, and that is not so good since
it polished source tree with kind of generated garbage.

Since Galera now supports build with Cmake, and it is not just a much
more faster then scons, but also allows to add the lib to our proj on
the source level, let's use this approach. Sources are fetched from
bundle dir, or from github, into our build (NOT source) tree. Then it
is configured and added to our whole build, also using wsrep placed
inside our sources. The whole project then builds and produces both
binaries of manticore and Galera plugin.

It also removes config params as 'WITH_REPLICATION', since now
replication is always 'on' by default on linux platform and actually
depends only from presence of galera library (i.e. - no library = no
replication; no need to reconfigure and recompile to achieve this)