Behave more friendly to long HTTP GET requests
Let's not hardly limit quantity of data received over HTTP session to
initial 4000, but instead grow slowly to adopt more data up to
max_packet_size. That allows to run long requests via GET method
(i.e. when it is placed in HTTP header, not in HTTP body).
It closes #800