Fix ub in russian utf8 stemmer
It checked end of a word by testing for zero last symbol. However
symbol was implicitly 2-bytes wide, but caller set to zero only 1-byte character. So such checking it for zero was
accidentally success only because of initially zero-filled buffer.

In worst case checking failed, and so the whole buffer is checked
until the end, and it is really ub since nobody guarantee that
2-bytes zero value will met before all available memory chunk
finishes (so, it may cause crash with SIGSEGV then).

That closes #814