Fix updates dword order
It is noticed that saving int64 into vec or DWORDs as aliasing
(union) it into 2 DWORDS and save first, then second; and also
saving (shrinking) DWORDs into int64 as placing the value, and then
placing 0 actually hadrware-depended on low-endian byte order. In
case of BE for this situation we confuse hi and low part of
bigint.

Also it remove AttributePacker_MVA32_c as redundant dupe, and reuse
AttributePacker_MVA_Update_T renamed to AttributePacker_MVA_T instead.