Add mmap readers for docs and hits
Fast experiment month ago revealed that accessing doclists and
hitlists using memory-mappings often makes searching significantly
faster. However in some cases it, in opposite, works slower, and
that is why direct drop-in replacement is not desirable. So, instead
of just replacing now we provide both methods, so that you can
choose the best approach in every case. Necessary params are
documented - refer to changed description of 'read_buffer' param,
and also to additional 'read_buffer_docs' and 'read_buffer_hits'.
That closes #753