Correct confusing clauses in the docs
We really use not 2 syscalls (seek + read), but 1 pread.
Also, warming index with prereading is no way guarantees it is fully loaded and keeped in the memory - such is possible only when enough free RAM available, and our 'massage' is just a way to tell the OS that we need to access this data fastest way.