Make config loading oneshot
Usually we read (or even exec) config file twice: first to collect
meta-info (like CRC32, timestamps, etc), and then to actually load and
parse content. Let's 1-st pass for checking also caches the content, so
that next loading just uses it.

It excludes re-executing of she-bang configs (so that hard scripts will
be run only once). And also it relaxes restrictions to file-object of
config, since now it opened once, read from beginning to end and closed.
It makes possible for experienced user to provide any readable things
as 'pipe' or 'socket' as the source of config (like /dev/stdin, etc.)