Refactor rotation
First, extract global 'disabled' indexes accumulator, since it became
missused (it is good to have it in main thread, but when it soaks into
replication and became MT, it need more deep testing to work well).
Then move rotation to separate task instead of thread.
It maybe solves #925, closes #926 and finally closes #697