Introduce TaskManager
It is internal scheduling thread and thread-pool for executing tasks.
It encapsulates oneshot jobs which need to be done in separate threads,
like ping, flush, rotate, flush uservars, etc.

Task is named couple of functions with payload. One performs the job,
second used to dispose payload is job has to be aborted. Task has to be
registered in TaskManager, and in other words called 'task flavour'.

Apart the name and actors, Task also has limits for num of workers and
size of queue.

With register task you may run or schedule a job, which is tuple of Task
+ payload, which will be executed by TaskManager. After performing the
job is released (so, it is always oneshot).

The job will be executed either by scheduler thread, either by another
free worker thread (depends from limit for workers in task flavour).

Single-threaded tasks assumes to perform light work without blocking.
For them limit of workers has to be zero. They will be performed by
scheduling thread, without creating or invoking any workers form thread
pool.

Heavy tasks need non-zero limit of workers, and it means, namely, how
many jobs of such task may be run in parallel. Say, task with workers=1
will be still single-threaded, but will be performed not by sheduling,
but by dedicated thread in thread pool (which will be started, if
necessary).

Every task has queue of jobs, from which workers takes them and execute.
Size of the queue also may be limited, and this way you can avoid
flooding the queue. So, for 'global' tasks like binlog flushing such
limit may has a sense (no need to stuck 50 flushes in a serie; just one
executing and one in queue enough, so setting the limit to 1 is good).
However another task like pinging a host is not so good here, since
we ping different hosts, and each is different job, so assing a global
limit for them is not practical.