Fix race in netloop reporting
When scheduling several network tasks we first added them to queue, and
then provide the num of queues to observer, having in mind that queue
is not run, and so, it is safe to gap between tasks and reporting.
But really queue may be in run by timeout, or even kicked by another
thread, so it became real race. So, the flow of executiong is rearranged
to avoid the race.

It resolves #915