Add tiny ports of boost::event and boost::optional
Event contains both 'oneshot' and 'broadcast' setting (in opposite to
our bicycle), and also avoids falling into syscalls when possible.

Optional is at all useful as both temporary storage of a value, and also
as exactly intended - to have or not have a value (instead of strange
error code or nullptr returning).