sphCheckQueryHeight -> ConsiderStack
It returns 3 flavours of values:
-1: current stack is enough, may run query as is.
0: no way to run the query; need to fire an error.
positive: size of stack to allocate.

It is assumed to be run in coro context. For positive we will run
dedicated coro with dedicated stack in order to finish execution.

The function replaces sphCheckQueryHeight in following contexts:
- CSphindex_VLN::MultiQuery
- CSphindex_VLN::MultiQueryEx
- RtIndex_c::MultiQuery
- SnippetBuilder_c::Impl_c::MarkHits

That restores work of test 144 (test 386 still fail)