Fix broken master-agent on Mac OS
On usual net flow we first connect, then send data.
On TFO we sent to non-connected socket, and TFO performs connection
and sending of the first chunk of data automatically.
Working async, we first expect error 'connection not established', which
is ok for non-connected socket. Then we poll the state, and when it
becames ready, and we could send some data to the socket, we
consider such connection 'established', and then the only error we expect
in the normal async flow is 'EAGAIN'.

However, Mac OS behaves quite tricky in such scenario.
First sending reported as successful, and reports 2048 bytes sent. So,
we begin consider this connection as 'established'.
However immediate sending next chunk fails with unexpected error
'connection is not established'. Since it is NOT 'EAGAIN', we consider
it as real error.

In reality first sending just copies 2048 bytes to net buffer and begins
to establish real connection.
Second turn meets full buffer and not yet established connection,
so it fails.

Let's silently consume data sent in very first chunk
without switching to 'established'.

This commit is safe to be backported to master branch.
That fixes #1407