Fix https port behavior
Connection with unsecured http to that port will be rejected with
error code 400.
Connection with remote agent to that port will be rejected with error
message 'Binary API request was sent to HTTPS port'.

Connection with https to any appropriate port (https and multiprotocol)
will be immediately dropped if ssl env is not valid (no lib, no certs,
whatever)

That closes #1552