Count deleted documents
That is for 'show index status'. For plain index that is just number of
killed documents. For rt that is total number over all chunks.
Value may be used in general to consider a chunk to be targeted for
optimization (say, having disk chunk for 500M of data, and killed rate
for 90% may mean, that actually it has just 50M of data, and the rest is
 killed and may be junked).