Remove age tag from rt RAM segments
Originally it was added to keep segments order on merge. That was need
since segments had kill-lists to be applied to older segments.
However value of the tag as static (i.e. global) value over instance
implied that all rt indexes have to live on one instance and can't be
moved and mixed with rt indexes on another instances (otherwise rt index
 with lower value of tag being loaded on the instance having indexes
 with higher such tag would break the order for another indexes).

In current implementation kill-lists are no more exposed among segments,
 so merging became symmetric and keeping order is no more necessary.
 However if some aspects are still depends from the order, this commit
 can be safely rolled back as whole.