Implement disk chunk dropping
For now it works only when optimizing pass found smallest disk chunk to
be fully abandoned (i.e. all documents are killed). In that case it is
no sense to perform merging of such chunk, and it can be simple dropped
away.

That works for progressive merge and override cutoff behaviour, i.e. if
it is asked to leave, say, 10 chunks, but we recognize that some of them
 are empty - all empties will be dropped despite of cutoff limit.

Manual dropping here implemented as merging into chunk '-1', as
`debug merge IDX 20 -1` - will drop disk chunk 20.

That closes #1641