Implement runtime lightweight indexes schema
Instead of full data for oneshot indexes it provides schema (i.e.
returns something similar to 'describe index'. Schema itself is
something as 'id, data, type', and outputed data is just names and types
of columns. Actual data is skipped and usually not generated.
That is useful, say, to recognize why your 'where foo<100' filter
doesn't work, as it might reveal, say, that 'foo' is string and so,
can't be compared with 1000.