Refactor net events lifetime
Use refcounting instead of ad-hoc suggestions. Move picking new events
before after processing fired events.

When connection processed out of net-loop and finished, socked
destroyed and usually everything works ok. But in internal polling
structures socked is still referenced and will be actually removed on
the next tick - recognizing that it is actually destroyed (have no link
to netloop, have negative timeout). Race is in the fact that such
checking performed over destroyed socket, so that memory where it was
placed might be already occupied by other structures, so checking value
of the link and timeout actually returns garbage (something belonging to
another object already). Using refcounting socket now closed and
released, then netloop in it's tick recognizes it and also removes and
releases that connection which is not destroyed.

That should fix #1745