fix to cmake 3.19
That is done because new CLion is shipped with very new cmake, which
became suspicious to our codebase.

Prominent change (however not very big influent to us) is rejecting
working with old scripts declaring min version 2.8 - now 2.8.12 is the
minimum required by cmake.

Another change was breaking of test names which was entered in 3.18 and
then fixed after many complains. However now it requires a policy to be
specified.