fix thread info registration
It is revealed that on fast codepath daemon sometimes crashes during
startup. As globally we have just about 4-6 infos which have to be
registered just once on startup, and that is usually happens not all at
the same time, the order wasn't managed in flavour of possible
concurrency. Practice revealed that it has to be managed this way,
however. This crash doesn't have separate issue opened, however it is
fixed here.