Constrain optimizations
That is to make expr tree friendly for back-walking.
Constrain implies, that (1) child elems of a node are ALWAYS before the
node itself (so that creating expression they're will be already created
when turn of the parent came), and (2) optimized nodes specially marked
(for now - set token to '-1'). That is to avoid create garbage.