Runaround to boost searching bug. Ad-hoc!
The only purpose of this ad-hoc is to runaround boost bug, reasoned by
our own bug which is not seamlessly-fixable right now.

We use self-built boost 1.72.0 built with option runtime-link=static

Some (all) systems provide pre-built boost made without this option.

To achieve both we find boost, first, with Boost_USE_STATIC_RUNTIME,
then repeat without this var.

Unfortunately Boost configurator has own bug, so that after first try it
 already defines imported library even if it actually 'not found'. And
 later it rejects to repead searching as target is already defined.

It is no way in cmake to remove previously added target.

So, for now we try to configure aside project, the only purpose of it is
 to realize whether we deal with static-runtime boost or not - and then
 make the only 'find package' with right flags.

It would be perfect, if we just rebuild all build dockers without this
flag, making all builds consistent. However if we do it right now, then
all previous builds will be broken, since they are, in opposite, expect
boost with this flag set, and they will fail to build because of this
very bug in the boost.

So, let's plan after this commit (and m.b. couple of releases) to
finally rebuild all dockers with non-static boost, then just remove this
 stuff with strange aside project ping-pong, finishing with
 onboard-available boost without static runtime (i.e., 'default' on all
 the systems).