Accurate counting of autoevent on Windows
counter is protected by mutex on pthreads, but not so on Windows. It seems that because of such race it can sometime become negative and fall into long (2G iterations) loop. (if that is true, that mean that origin which is rewritten from boost-asio is also defective same way). That might be origin of github#578.