implement immutable rt-index management
Ram segments and disk chunks now stored as const refcounted arrays of
const refcounted pointers. Any reading op, like search, takes snapshot
of that data and releases it when finished. Any changing op, like
insert or optimize, makes copy of original array, change it as desired,
and finally swap to original arrays, so that whole change applied
in one-shot, with very short lock need only to change one pointer to new
 one.

Write-through operations, like applying kill-lists or updates, which now
 can't be performed in full isolation from save/merge are collected and
 then applied to result of save/merge.

Global flags 'disable optimize' and 'disable save' for now are switched
off. They NEED to be reviewed and then restored/refactored; before that
happened the functions depend on them should be considered as
broken, and so, that commit is NOT production-ready