Review index save disabling
DISCARD state if for indexes which will be totally discarded, so
completion of the ops is not necessary.
DISABLED actually disables saving, but allow saving attributes. That is
used before replication.
Also DISABLED state actively prohibits attr updates. That is necessary,
because otherwise index will be replicated in non-consistent state, as
updates are applied immediately via mappings.
In the case when update is called during SST replication to donor node,
error 'index is locked now, try again later' will be issued. Yes, that
is not seamless, but however better, then allowing corrupted index sent
to remote node. That is related to #2285