implement another optimize guarding
It was mutex before. Now it is atomic counter of 'how many optimize
tasks works right now', + boolean flag 'disable optimize', + event
'number of optimize works decreased'.

Some operations as altering index, and also index truncate implies that
current optimize step (if any) should be canceled, since it's result is
not necessary, and we don't have to spend time on it. The save was about
attaching disk chunk, but actually it doesn't harm optimization.

Also, on sst replication we lock file state, so that optimizing also
should not work during it, and that is even more critical.

Using mutex there generally solved the problem legacy way, that is -
holding same mutex by optimize pass and another commands imply that they
can't work together. But as a side effect these operations themselves
became linked to the mutex and mutually depends from it. Also,
long-term mutex holded during whole long-running job is not at all good
for coroutined tasks with context-switching, as in this case task
waiting for the mutex not rescheduled, but effectively pauses whole
OS thread.

Now optimize just checks 'disable optimize' before start, and also
several times during the process (as it was before). Also, it increases
number of active optimizes (index-scoped) on enter, and decreases it on
exit. Also it sets event 'number decreased' on exit. That is - now it is
possible to run simultaneously >1 optimize pass on one and same index
(however chunk selection should be aware, pairs of chunks should not
intersect).

Tasks requiring disabled optimize first, set the bool 'disable
optimize', and then waits in loop until number of active optimize passes
is abandoned. Loop is managed by event (so, it is NOT hot spin-loop).

And the final thing - as we had coroutined event designed only to one
waiters, but here more than one op can wait for optimize changes, we
implemented multi-event for that. It holds queue of operations and calls
all of them on fire.

That fixes #2281