Apply rtsegments limiter
That is: Commit is deffered until N of free segments is < 32. Free
segments is ones which are not participating in current segments merging
or saving as disk chunk. So, if a save op occupies 32 chunks - they are
not in the count, as they'll be finally released, and look only into
current N of free chunks.