Adaptive mem limit
Our default legacy is to collect up to rt_mem_limit data, then start
saving to disk, and at the same time still collect up to 10% of
rt_mem_limit in advance (aka double-buffer). If that limit also
exchaused, insertion stalls.

Adaptive limit built on the fact that we have auto-optimize now, so that
is not very big deal if disk chunks will not fully adopt rt_mem_limit.
That is: now we collect up to 50% of rt_mem_limit and save disk chunk.
On the end of save we look to the rate (how many we save, how many new
data arrived during the save) and recalculate initial rate. That is, if
we saved 90M, and newly arrived 10M - rate is 90%, so that we then can
collect up to 90% of rt_mem_limit before start flushing. That is -
higher speed of insertion, lower mem limit rate. Values are selected
automatically from 33.3% to 95%.