more robust segment merge tracking
That is - if during merge we request another one, it is rejected with
thought that at the end of merge we anyway will reschedule. However,
that is not always true, as merge can finish with flush, or can finish
without any changes (say, it was caused by 'delete' txn, which didn't
totally kill one of segments, and so, no op performed). Now incoming
merge request memorised, and if a merge currently active, it will check
memorized value at the end, and restart, if necessary.