kick segment merging
That is necessary if insertion is impossible due to N of segments, but
current or previous merges are finished without changing the N (say, it
was 'kill' merge, or restart with N without any inserts yet). Then we
kick invoke merge op, waiting that it will decrease N of segments, and
so, we could then continue insertion.

Test case: start daemon with the index containing >=32 segments, and try
 to insert anything.

That fixes #2376