mem_limit_rate report
Displayed in 'show index status', indicates part of mem_limit to be
achieved before RAM->DISK flush will happen.
(initialy 50%, then may grow up to 90% on non-intensive inserts, or
decrease down to 33.33% on very high insert rate. Rationale: part of
mem_limit is reserved to accept inserts when ram chunk is already
flushing. It should be enough to not be overloaded until flushing is
finished.)