Unify objdump invocation
That is about build/config

objdump is used to extract symbols on mac os
We call it with flags `-macho -dylib-id`, that works for mac os.
However another objdump from clang toolchain doesn't recognize them,
and it needs `--macho --dylib-id`.

Probe revealed, that last is also acceptable for apple objdump, so it
works both for native apple, and for clang toolchain.