process 100-continue
Manticore currently doesn't support HTTP header Expect: 100-continue
which some clients (e.g. curl) send and then wait for some time
before sending POST data.

The task is to add support for Expect: 100-continue the way that
Manticore always confirms immediately the client can continue.

Additional legend

Original implementation was build solely on the old thread-pool and the
route was made blob-oriented.
That is:
1. read whole data from socket,
2. parse it.

By "whole data" it was naturally everything, i.e. http header and http
body. Distinguish was done later by http parser.

A bit later it was realized, that for large packages reallocation of
network buffer makes hit to cpu and memory consumption, and as ad-hoc it
was added 'pre-parser': that is totally another piece of code, apart
from http_parser, written with only purpose: trace incoming data and
figure out 'Content-Length' field from it. Having this number known, it
then allocate whole buffer and so, avoids reallocations.

As a side effect, if there is no content-length (say, it is 'chunked',
or 'connection: close') variation, we can't process such requests and
dropped them (yes, that is sorry limitation done exclusively because of
this ad-hoc).

True parsing, btw, was still done by http_parser (that is 3-rd party
lib), and that actually re-parse again which was already feed to
'pre-parser'.

Also, we pick only some parts of the header with this parser (that is -
url and fields). Body was never processed, as we expect only fixed solid
blob with known content-length, which was directly passed to our engine.

However, closer look to http_parser reveals that it is true stateful
parser, which can be feed by chunks of data with any size. That is: we
can receive available chunk from connection, parse it, then receive next
one, etc.

That means, that http_parser can itself solve task of extracting
'content-length' field; we don't need any kind of 'pre-parser' for it.

Also it can signal when whole header is read. At this point we can
perform additional actions, like answer to 100-continue request.

Also it can signal when whole body is read, and it supports both
'content-length' for solid encoding, and 'termination chunk' for chunked

These opportunities are good reason to remove 'pre-parser' and use
solely http_parser. That makes code simpler, and also on the base of
co-routines it allow to process incoming data by pieces (separately
'header' and then 'body')

That fixes #2601
That fixes #2676