zero-copy http
When parsing http requests we make copy of incoming data, because
parsing implies intrusion (that is - we place \0-terminator in some
places in order to make json parser happy).

However actually we don't need that arrived data untouched, since it's
only purpose is to be parsed and processed. So, let's use directly
incoming buffer and avoid copying it.

Some corner cases should be traced:

1. URL-decoding. We sometimes need both url-decoded and raw line.
However these chunks are tiny, and ok to copy.
2. Text of data to be exposed for 'show threads'. Here is also, that is
no reason to keep huge amount of data; but some arbitrary chunk, say,
100 chars, should be enough. That is ok to copy such tiny chunk.

What is done here:

Perform parsing of incoming data directly in network buffer instead of
redundand copying it.

Note: for bulk we acquire up to first 100 bytes into task info (that is
line exposed in the 'show threads').

That closes #2677