new task/sched system
Execution via systhreads was added years ago with old thread-pool. That
is: all async tasks like rotation, pings, malloc-trim, etc. was executed
 as 'tasks' in dedicated system threads. However, when new thread pool
 came with coro execuction, that tasks were modified to be run in main
 thread pool instead, and system threads became just 'proxy jumpers',
 which issue real job into main thread pool and then wait until it
 finished.

Here this 'proxy jumpers' are finally eliminated. That is - no more
'system threads', every async task just fired into main thread poll
directly. The difference between usual plain coroutine and 'the task' is
just addition ticket with statictic and limiters (that is - name of the
task, time spend to task, N of runs, etc.)

Another thing is regarded 'mini-timer' and 'task scheduler'. Both are
about waiting. Mini-timer was introduced to avoid bulk call of
GetTimeOfDay(), which revealed to be quite heavy and significantly
slowed queries with time limit. Instead, dedicated thread was calling
such GetTimeOfDay() every 1mS and share result to all threads. That
decreased hit of calls, however calling the func every 1mS was still
compromise, 'better than worse'. Actual task of mini-timer is to
mark some moment in the future, so that we can check whether the moment
is achieved, or not yet (ticking every 1mS actually is not necessary).
Say, if we want to run something with timeout 1S - we can engage the
timer on 1S, and when it is achieved, tick only once, signalling that
period is exceeded. On main executor, then, we just check the boolean
'exceeded or not', and don't need intermediate ticks. So, actual
engaging is literaly 'mark signal in given time'.

Task scheduler, in turn, was part of
systhreads to schedule execution of async tasks. That was reasonable
with dedicated system threads. With eliminating of them it became much
less specific, and now just fires given handler (function) at given
moment of time into main thread pool (with system threads it also
performed very specific tasks like starting system threads and manage
their lifetime).

That is common point of scheduler and mini-timer - both just run
something in given moment (scheduler issues actual task, mini-timer just
 'mark' some period exceeded). So, their both are unified into one
 scheduler which serves both for timeouts and for deferred execution of
 tasks.