fix deb dependencies
More granular dependencies for packages. That is, 'server', 'icudata'
and 'dev' actually don't deal with 'libmysqlclient', and so, should not
recommend it.

'manticore-icudata' on debian is actually 'manticore-icudata-65l'

All dependencies investigated via shlideps are made explicit. That,
first make building a bit faster, and also allows cross-building, as
host system in that case doesn't has enough stuff to introspect shared
libs on alien target system.