change workers_active/tasks value
'workers_active' is field in 'show status', the same 'Tasks' is value
in oneline 'status'.

It was overall counter of all created job descriptors, however that is
not so useful, as several jobs may be chained together (for example,
encrypted compressed connect via mysql has 3 descriptors - one for
'mysql', one for 'compressed', one for 'encrypted'. But that is 1 job,
not 3).

Now it summarises only connections and tasks. Tasks are global (like
periodic flush), and also rise on parallel execution (query parallelized
 into 4 threads will add 3 tasks to currently running connection).