implement doc extract
Run via indextool, as `indextool --docextract IDX DOCID`, for example,
indextool --docextract index2 8042572

It runs usual index check pass of whole dictionary/docs/hits, and
collects all the words and hits belonging to requested document.

Then all of the words are placed in the order according to their fields
and positions, and result is printed, grouping by field.

That fixes #2891