change CSphWriter/autofile behaviour
autofile with 'temporary' flag can be set to 'persistent' - that is,
non-temporary. You can start with 'temporary' file, and if final result
is need to be stored - reset temporarity with 'SetPersistent'.

CSphWriter implies explicit call to CloseFile() (if not owned),
otherwise file will be unlinked on destructor. That is - all unfinished
operations interrupted before file is closed imply automatic clean-up
will be performed.

DeleteOnFail_c was removed because of these changes as redundant.
Explicit CloseFile() in some place were added.