fix inter-thread socket race
It happens in some specific cases when we're going to 'release' socket
and don't poll it anymore. Since socker management and net-loop works
from different threads, it can happen, that on releasing socket is
destroyed, and then net-loop thread faces with garbage because of such
use-after-delete.

That should fix #3201, and may be also related to #3069