use guarded stack
Define env MANTICORE_GUARDED_STACK in order to force the daemon to use
guarded stack instead of plain one.

Guarded stack uses guard page on the top, so that in case of too deep
recursion, page will be touched, and it will cause segmentation
violation. That helps to early localize such abnormal flow, because
otherwise it causes undefined behaviour and most probably crash, but
much later, when we can't figure out what is the reason