uniform reply to bulk
Bulk endpoint (which takes ndjson stream) now report same kind of jsons
both in case of success and in case of error (use HTTP return code to
distinguish).

Reply includes usual stuff ('items', 'errors'), and also 3 extra fields:
"current_line" - number of last line in stream (remember, ndjson is list
 sequence of lines, separated by cr or lf. So, each document is
 literally a line). "skipped_lines" - number of non-commited lines. That
is, if you have several bulks in a stream - say, 10 inserts to one
index, then 10 inserts to another, then 10 inserts to third - and on
5-th line of the last txt something bad happened, like invalid json -
you will see skipped_lines to be 5, and current_line to be 25. That
means, to repair you need to review and try again only last 5 lines
(docs) and don't need to send whole stream from the beginning, as first
20 are successfully processed to the moment. The last field - "error" -
is just a text of error. It is filled, if 'errors' is true, and usually
empty otherwise.

That should fix #3348