Eliminate racing closed connection in master-agent
Let's behave for kqueue route same as for epoll - that is, divide
appending of new events from general polling loop. With previous
behavior, when we postpone changes to the one single kqueue call - it
was race, when such postponed connection was deleted after scheduling,
and then on actual call wrong callbacks, etc. was called over this
deleted thing.

Without postponing we sure, we process alive connection, and if it is
then closed - both epoll and kqueue automatically remove it and so wrong
 callback will not be activated.

CHANGELOG: eliminated crash when master can't connect to agent on
kqueue-driven systems (freebsd, macos, etc.)

That closes #3382