make timeout message more clear
Actually, when we report 'connect timed out' for master-agent
communication, it may mean not just 'connect', but full query round,
including connect, query, and waiting for response. That is because of
TFO we may think about just 'connect' stage, when actually whole query
is already sent in syn package, and 'query' timeout is already in game.

That fixes #3171