generate shadow model in CI
When running ubertests, we compare current result with stored model and
generate textual human-readable report. However, if a tiny difference
revealed, test will fail, and then need manual intervention to renew
model, if it need to be changed.

Shadow model is file 'report.bin', placed in test folder across with
textual 'report.txt'. It includes generated model based on the test run
which is described in report.txt.

So, if test fails, and from diff it is obvious, that new behavior is one
 that should actually be - you can just download that shadow model
 from CI artifacts, and then replace 'model.bin' in the test with this
 'report.bin' file.