upgrade mysql capability of ignoring EOF
EOF packet is deprecated and replaced with OK packet. According to mysql
docs, that works starting from Mysql 5.7.5, and managed by special bit
flag in capabilities. So, if client is old and doesn't support such
capability - we will work with it in old fashion, with EOF package. For
modern client behaviour changes in 2 places. First, we send OK instead
of EOF. Second, we don't send anything between header and body of
resultset (in old fashion that is one more EOF there).