aside Galera build
when build with -dPACK=1 we check, if env PACK_GALERA is set to 1.
In this case galera packages will be created and placed near manticore
packages, so that all artifacts are in place.

It is suggested to set PACK_GALERA once, since it is no sense to rebuild
 it again and again; oneshot build should be enough.

That implements stuff need for #416, but extra steps maybe necessary for
 CI.