implement out-of-order updates
That is, request for update during running optimize pass will be
injected into running merge instead of waiting the last for completed.

That eliminates blocking of updates by optimize.
However, that is not the final commit; some checks also need to be
performed first.

That is related to #2361 (2-nd subtask)