fix unfreeze<=>update deadlock
Trying to update a blob attr in frozen index leads to condition, when
index record is exclusively locked, and update routine waits the index
to be unfrozen. In this case trying to unfreeze by name will also try
to lock index record, but that try will cause the deadlock.

related to manticoresearch-backup#116