wait enabled save before update
When update happens, it first acquire shared or
exclusive lock on table, and then start action.
If index is frozen, it will wait until it is released.
Unfortunately, with exclusive lock all the other operations will also
stuck here.

Here we ensure index is not frozen before update, and also ensure it is
NOT frozen during update. Necessary shared/exclusive lock is taken with
that ensure.

This also revert 9987798a as non-necessary.
Related to #2737